Shunt AccessStatus.apply() into the Sierra transformer
That's the only place it's used now -- this logic is complicated enough,
we try to limit it to per-source rather than having overlapping rules
for different sources.

It's one less thing to have compiling in internal_model.
