+++ pipeline/transformer/transformer_calm/src/main/scala/weco/pipeline/transformer/calm/CalmTransformer.scala
@@ -100,19 +100,6 @@ object CalmTransformer
                   List(knownErrToUntransformableReason(knownErr))
               )
             )
-          case unknownStatus: UnknownAccessStatus =>
-            warn(
-              s"${record.id}: unknown access status: ${unknownStatus.getMessage}"
-            )
-            Right(
-              Work.Invisible[Source](
-                state = Source(sourceIdentifier(record), record.retrievedAt),
-                version = version,
-                data = WorkData(),
-                invisibilityReasons =
-                  List(InvalidValueInSourceField("Calm:AccessStatus"))
-              )
-            )
           case err: Exception => Left(err)
         }
     }
